requirements.txt
cryptography==36.0.1
dash==2.0.0
dash-bootstrap-components==1.0.2
dash-core-components==2.0.0
dash-daq==0.5.0
dash-html-components==2.0.0
dash-table==5.0.0
dask==2021.12.0
Flask==2.0.2
Flask-Compress==1.10.1
Flask-Cors==3.0.10
Flask-Excel==0.0.7
numpy==1.21.5
openpyxl==3.0.10
pandas==1.4.4
plotly==5.5.0
python-dotenv==0.19.2
shioaji==1.1.0
XlsxWriter==3.0.2
lxml==4.8.0